to remove the extra layer of folder depth
