+++ Source/Client/Features/AgldRate/AgldGetRateAction.cs
@@ -1,0 +1,10 @@
+﻿namespace Client.Features.AgldRate
+{
+  using MediatR;
+
+  public class AgldGetRateAction : IRequest<AgldRateState>
+  {
+    public decimal AgldRate { get; set; }
+
+  }
+}
