--- Source/Client/Features/Conversion/AgldRate/AgldGetRateAction.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using MediatR;
-using Shared.Features.Conversion;
-
-namespace Client.Features.Conversion.AgldRate
-{
-  public class AgldGetRateAction : IRequest<AgldRateState>
-  {
-    public decimal AgldRate { get; set; }
-
-    
-  }
-}
